+++ graphics/graphics.hpp
@@ -18,14 +18,14 @@ class pGraphics {
 
     pColor white, black;
     
-    void drawPolygon(std::pair<int, int> p1, std::pair<int, int> p2, std::pair<int, int> p3, std::pair<int, int> p4, pColor color);
-    void drawRect(std::pair<int, int> pos, std::pair<int, int> size, pColor color);
-    void drawFilledCircle(std::pair<int, int> pos, double r, pColor color);
-    void drawText(std::pair<int, int> pos, void *font, const char* str, pColor color);
-    void drawTriangle(std::pair<int, int> point1, std::pair<int, int> point2, std::pair<int, int> point3, pColor color);
-    bool mouseInRegion(std::pair<int, int> mousePointer, std::pair<int, int> pos, std::pair<int, int> size);
+    void drawPolygon(std::pair<double, double> p1, std::pair<double, double> p2, std::pair<double, double> p3, std::pair<double, double> p4, pColor color);
+    void drawRect(std::pair<double, double> pos, std::pair<double, double> size, pColor color);
+    void drawFilledCircle(std::pair<double, double> pos, double r, pColor color);
+    void drawText(std::pair<double, double> pos, void *font, const char* str, pColor color);
+    void drawTriangle(std::pair<double, double> point1, std::pair<double, double> point2, std::pair<double, double> point3, pColor color);
+    bool mouseInRegion(std::pair<double, double> mousePointer, std::pair<double, double> pos, std::pair<double, double> size);
 
-    std::pair<int, int> getTextSize(const char* str, void* font);
+    std::pair<double, double> getTextSize(const char* str, void* font);
 
     pColor createNewColor(int r, int g, int b, int a = 255);
 };
