+++ interface/interface.hpp
@@ -7,10 +7,10 @@
 class pInterface {
     private:
     void(*render)(void);
-    void drawSquare(std::pair<int, int> pos, pPieceColor color);
-    void renderSquare(std::pair<int, int> pos, pPieceColor color, std::vector<std::pair<int, int>> &posMax, bool show = true);
+    void drawSquare(std::pair<double, double> pos, pPieceColor color);
+    void renderSquare(std::pair<double, double> pos, pPieceColor color, std::vector<std::pair<double, double>> &posMax, bool show = true);
     void updateMaxPos(pPiece piece);
-    void computeHitbox(pPiece &piece, std::pair<int, int> pos);
+    void computeHitbox(pPiece &piece, std::pair<double, double> pos);
     void checkLevelOrScoreIncrease();
     public:
     pInterface(void(*pRender)(void));
@@ -31,7 +31,7 @@ class pInterface {
     void onKeyPress(unsigned char key);
 
     void automaticMove();
-    std::pair<int, int> predictFallingPosition();
+    std::pair<double, double> predictFallingPosition();
     void delRow(int row);
 
     int genRandomNumber(int min, int max);
@@ -43,7 +43,7 @@ class pInterface {
     pPieceColor cyan, blue, yellow, green, purple, red, orange, block;
     pPieceColor nullCol;
     std::vector<pPiece> pieces;
-    std::vector<std::pair<int, int>> border;
+    std::vector<std::pair<double, double>> border;
 
     void spawnPiece();
 };
